# [FIX] web: speedup basic_model._updateRecordsData