Speedup `basic_model._updateRecordsData` for
many2many fields such as `fetchmail.server.message_ids`
and `mail_channel.channel_message_ids`.

Calling `_.findWhere(values)` inside two nested `_.each`
calls can be quite slow when the number of values
is big.

Remove the call to `_.findWhere` by first
creating a Map res_id -> data and then calling
Map.get to retrieve the data for a given res_id
when making a DataPoint

The issue this fix solves happens in v13, v14 and v15 (master wasn't tested).

#### speedup

Time reported (on Chrome) for loading the fetchmail.server tree view.
The search_read call is included.

| len(message_ids) | Before PR | After PR |
|:-----------------:|:----------:|:---------:|
|          5000         |   1.4s        |   291ms |
|        10 000        |    4.6.s      |   475ms |
|        20 000        |    18s        |   860ms |
|        200 000      |   26m       |      7s     |

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr