@robodoo 2021-12-08
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/81067).